+++ src/main/java/tavebalak/OTTify/common/constant/Role.java
@@ -1,4 +1,4 @@
-package tavebalak.OTTify.oauth.constant;
+package tavebalak.OTTify.common.constant;
 
 
 public enum Role {
