+++ src/main/java/tavebalak/OTTify/oauth/CustomOAuth2User.java
@@ -3,7 +3,7 @@ package tavebalak.OTTify.oauth;
 import lombok.Getter;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.oauth2.core.user.DefaultOAuth2User;
-import tavebalak.OTTify.oauth.constant.Role;
+import tavebalak.OTTify.common.constant.Role;
 
 import java.util.Collection;
 import java.util.Map;
