+++ src/main/java/tavebalak/OTTify/oauth/handler/OAuth2AuthenticationSuccessHandler.java
@@ -4,10 +4,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.authentication.AuthenticationSuccessHandler;
-import org.springframework.security.web.authentication.SimpleUrlAuthenticationSuccessHandler;
 import org.springframework.stereotype.Component;
 import tavebalak.OTTify.oauth.CustomOAuth2User;
-import tavebalak.OTTify.oauth.constant.Role;
+import tavebalak.OTTify.common.constant.Role;
 import tavebalak.OTTify.oauth.jwt.JwtService;
 import tavebalak.OTTify.oauth.redis.RefreshTokenRepository;
 import tavebalak.OTTify.user.repository.UserRepository;
