+++ src/main/java/tavebalak/OTTify/user/repository/UserRepository.java
@@ -2,7 +2,7 @@ package tavebalak.OTTify.user.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import tavebalak.OTTify.oauth.constant.SocialType;
+import tavebalak.OTTify.common.constant.SocialType;
 import tavebalak.OTTify.user.entity.User;
 
 import java.util.Optional;
